refactor: delete index.js
